AI_Code/ai_requirements.txt
#create a conda enviornment with python 3.7
#RUN THE BELOW COMMAND TO CREATE THE SAME:
# conda create -n snakeenv python=3.7   
# conda activate snakeenv

ultralytics
fuzzywuzzy
flake8 
pycodestyle_magic
inference_sdk
--find-links https://download.pytorch.org/whl/torch_stable.html
blis==0.7.4
catalogue==1.0.0
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
cymem==2.0.5
fastai==2.3.0
fastcore==1.3.20
fastprogress==1.0.0
idna==2.10
importlib-metadata==4.6.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
murmurhash==1.0.5
numpy==1.21.0
packaging==20.9
pandas==1.2.5
Pillow==8.2.0
plac==1.1.3
preshed==3.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
spacy==2.3.7
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
timm==0.4.5
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wincertstore==0.2
zipp==3.4.1
